requirements.txt
Flask
flask_sqlalchemy
flask_migrate
psycopg2
python-dotenv